Restore quantising to `i16`by default. If quantisation fails (due to overflow), an error message will be displayed, prompting the user to quantise manually. Training will be otherwise unaffected.

+++ src/trainer.rs
@@ -113,7 +113,7 @@ impl<T: InputType> Trainer<T> {
 
         self.optimiser.write_weights_to_host(&mut buf);
 
-        let mut qbuf = vec![0i32; size];
+        let mut qbuf = vec![0i16; size];
         let mut qiter = self.quantiser.iter().peekable();
         while let Some(&QuantiseInfo { val, start }) = qiter.next() {
             let end = if let Some(QuantiseInfo {
@@ -127,7 +127,7 @@ impl<T: InputType> Trainer<T> {
 
             for i in start..end {
                 let qf = (f64::from(val) * f64::from(buf[i])).trunc();
-                let q = qf as i32;
+                let q = qf as i16;
                 if f64::from(q) != qf {
                     println!("================= WARNING ================");
                     println!("   An error occured during quantisation:  ");
